requirements.txt
requests>=2.31.0
pandas>=2.0.0
tqdm>=4.49.0
urllib3~=2.2.1
scikit-learn
numpy
joblib
PyYAML
wandb
loguru
fire
aiohttp
lightgbm
skorch
plotly
matplotlib
seaborn
arfs

# eda
umap-learn

# SparCC
dask[complete]
numba
h5py
scipy


# dev
ruff
nbformat
pytest